Add sobol sampling
Refactored some more things. There are outstanding issues with sobol
indices.

+++ src/inputs/jointdistribution.jl
@@ -12,7 +12,7 @@ struct JointDistribution <: RandomUQInput
     end
 end
 
-function sample(jd::JointDistribution, n::Int64 = 1)
+function sample(jd::JointDistribution, n::Integer=1)
     u = sample(jd.copula, n)
 
     samples = DataFrame()
